refactor external providers tests
