+++ packages/web3-types/src/web3_base_provider.ts
@@ -28,6 +28,7 @@ import {
 	JsonRpcSubscriptionResult,
 } from './json_rpc_types';
 import { Web3APISpec, Web3APIMethod, Web3APIReturnType, Web3APIPayload } from './web3_api_types';
+import { Web3EthExecutionAPI } from './apis/web3_eth_execution_api';
 
 const symbol = Symbol.for('web3/base-provider');
 
@@ -166,7 +167,7 @@ export abstract class Web3BaseProvider<API extends Web3APISpec = EthExecutionAPI
 	public abstract reset(): void;
 }
 
-export type SupportedProviders<API extends Web3APISpec> =
+export type SupportedProviders<API extends Web3APISpec = Web3EthExecutionAPI> =
 	| EIP1193Provider<API>
 	| Web3BaseProvider<API>
 	| LegacyRequestProvider
