+++ packages/web3/test/integration/external-providers/hardhat.test.ts
@@ -1,0 +1,28 @@
+/*
+This file is part of web3.js.
+
+web3.js is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+web3.js is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License
+along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+// eslint-disable-next-line import/no-extraneous-dependencies
+import hardhat from 'hardhat';
+
+import { performBasicRpcCalls } from './helper';
+
+describe('compatibility with `hardhat` provider', () => {
+	it('should initialize Web3, get accounts & block number and send a transaction', async () => {
+		// use the hardhat provider for web3.js
+		await performBasicRpcCalls(hardhat.network.provider);
+	});
+});
